refactor: {XXX}SimpleSearchProfile Kotlin 파일 응집화
